# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Karigar Service Marketplace API",
        "description": "RESTful API for Karigar - A service marketplace connecting customers with skilled service providers (electricians, plumbers, carpenters, etc.)",
        "contact": {
            "name": "Karigar API Support",
            "email": "support@karigar.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Karigar API Server"
        }
    ],
    "paths": {
...
        "/api/requests": {
            "get": {
                "tags": [
                    "Service Requests"
                ],
                "summary": "Get all service requests",
                "description": "Retrieve service requests filtered by user role. Customers see their requests, providers see requests for them.",
                "operationId": "d8762545247b9ed012f00e992cbfa665",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "requested",
                                "confirmed",
                                "completed",
                                "cancelled"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Requests retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Service Requests"
                ],
                "summary": "Create new service request",
                "description": "Customer creates a booking request for a service provider",
                "operationId": "b78212afe2734cba4b079913cc3d3855",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "provider_id",
                                    "service_id",
                                    "requested_date",
                                    "requested_time",
                                    "address"
                                ],
                                "properties": {
                                    "provider_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "service_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "requested_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-25"
                                    },
                                    "requested_time": {
                                        "type": "string",
                                        "example": "10:00 AM"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Need electrical wiring repair"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St, New York, NY"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service request created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/requests/{id}": {
            "get": {
                "tags": [
                    "Service Requests"
                ],
                "summary": "Get request details",
                "description": "Get detailed information about a specific service request",
                "operationId": "6c4f2a70cb6d6ec470c9eff7499c9bce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request details retrieved"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/requests/{id}/accept": {
            "put": {
                "tags": [
                    "Service Requests"
                ],
                "summary": "Accept service request",
                "description": "Service provider accepts a requested booking",
                "operationId": "701b8c23be19ff43b3e0a55c675b036e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request accepted"
                    },
                    "400": {
                        "description": "Invalid status transition"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/requests/{id}/reject": {
            "put": {
                "tags": [
                    "Service Requests"
                ],
                "summary": "Reject service request",
                "description": "Service provider rejects a requested booking",
                "operationId": "e592214df4f818cb242463ca709da3f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Not available on requested date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request rejected"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/requests/{id}/reschedule": {
            "put": {
                "tags": [
                    "Service Requests"
                ],
                "summary": "Reschedule service request",
                "description": "Service provider reschedules a booking to a new date/time",
                "operationId": "ce82aeb78d7a120a8aceef832f3ea541",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "new_date",
                                    "new_time"
                                ],
                                "properties": {
                                    "new_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-26"
                                    },
                                    "new_time": {
                                        "type": "string",
                                        "example": "2:00 PM"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Previous booking conflict"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request rescheduled"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/requests/{id}/complete": {
            "put": {
                "tags": [
                    "Service Requests"
                ],
                "summary": "Mark request as completed",
                "description": "Service provider marks a confirmed request as completed",
                "operationId": "677f5867b91ad9425ee92282ad591f69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request marked as completed"
                    },
                    "400": {
                        "description": "Invalid status"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/requests/{id}/cancel": {
            "put": {
                "tags": [
                    "Service Requests"
                ],
                "summary": "Cancel service request",
                "description": "Customer or provider can cancel a requested/confirmed booking",
                "operationId": "91fad068c306105d9f231208b7f3218d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request cancelled"
                    },
                    "400": {
                        "description": "Invalid status"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}